fix: repair the #1331 merge — one generation of #1295, one engine, one bus
`main` has not compiled since #1331. Its branch merged `main` in and resolved
three conflicts by keeping both sides, which left two crates unparseable and a
shipped knob unreachable. Each fix takes the side `main` already had.

**#1295 had two generations, and both landed.** #1328 shipped the refined
design to `main` — `judge_evidence_demand`, gated on a tracked command so the
ask is only raised where it can be answered (#1211 §1). The #1331 branch still
carried its own earlier draft, `require_evidence_for_lone_judge_pass`. The
merge kept both config fields, both CLI knobs, and spliced both bodies into
`verify_candidate`, leaving a `revise_candidate(` call unterminated. The older
draft is removed: its const, its field, its CLI plumbing, and the two tests
that exercised it — `tests/judge_evidence_demand.rs` covers all four of their
cases plus the no-tracked-command one they never had.

The draft's TOML plumbing had *displaced* the shipped knob's, so
`pipeline_judge_evidence_demand` parsed as an unknown key, never reached
`PipelineConfig`, and was dropped by `AgentEngineConfig::overlay` — the same
silently-lower-scope-wins bug the `headless_scope_bypass` comment describes.
Renaming that plumbing restores it.

**`ServeConfig` lost a field initializer.** `extensions` (#1298) was spliced
into the body of `with_sub_agents` (#1297), which then returned nothing.

**`run_session` built two engines.** One over `&tools` with the hook bus and
the calibration map (#1298), one over the delegating `tool_view` that makes
sub-agents reachable (#1297) — with neither attachment. They are now one
engine over `tool_view`, carrying both. The duplicated `drive_turn` that
followed the goal-run `match` is gone, and the bus reaches every round of a
goal run and every sub-agent's tool executor: an operator policy that can
refuse a tool call must not be escapable by delegating it.

Also: the `SessionSpec` literals in both serve suites carry the other side's
fields, `api_sources()` counts the two files it gained, `judge_evidence_demand`
is switched off in the OOM scenario the way its siblings already do it, and
`routes.rs` — 1521 lines once both sides' routes landed — moves `/v1/sessions`
into a submodule, with the two now-obsolete baseline entries retired.
