+++ scripts/file-size-baseline.txt
@@ -17,21 +17,19 @@
 1615 bench/terminal_bench_analysis/tests/test_tb21_evidence_contract.py
 2370 stella-cli/src/agent.rs
 1748 stella-cli/src/agent_tests.rs
-1676 stella-cli/src/candidate_ws.rs
-4682 stella-cli/src/command_deck.rs
+1576 stella-cli/src/candidate_ws.rs
+4582 stella-cli/src/command_deck.rs
 1516 stella-cli/src/fleet_cmd.rs
-1526 stella-cli/src/memory/tests.rs
 2129 stella-core/src/bus.rs
 2764 stella-core/src/driver.rs
 3661 stella-core/src/driver/tests.rs
-1632 stella-graph/src/store.rs
 1677 stella-model/src/anthropic/tests.rs
 2072 stella-model/src/openai.rs
 1572 stella-model/src/zai.rs
 1843 stella-model/src/zai/tests.rs
-4697 stella-pipeline/src/pipeline.rs
-4616 stella-pipeline/src/pipeline/tests.rs
-4723 stella-protocol/src/event.rs
+3666 stella-pipeline/src/pipeline.rs
+2621 stella-pipeline/src/pipeline/tests.rs
+2723 stella-protocol/src/event.rs
 2078 stella-store/src/lib.rs
 2261 stella-store/src/tests.rs
 1917 stella-store/src/usage.rs
