+++ stella-cli/src/settings/unknown.rs
@@ -96,7 +96,7 @@ pub(crate) const ENGINE_ROOT_FIELDS: &[&str] = &[
     "headless_scope_bypass",
     "pipeline_max_revisions",
     "pipeline_candidates",
-    "pipeline_evidence_for_lone_judge_pass",
+    "pipeline_judge_evidence_demand",
     "pipeline_require_diff_coverage",
     "model_timeout_secs",
     "compaction_budget_tokens",
@@ -192,7 +192,7 @@ const TOML_AGENTS_FIELDS: &[&str] = &[
     "headless_scope_bypass",
     "pipeline_max_revisions",
     "pipeline_candidates",
-    "pipeline_evidence_for_lone_judge_pass",
+    "pipeline_judge_evidence_demand",
     "pipeline_require_diff_coverage",
     "default",
     "worker",
