+++ stella-parity/src/lib.rs
@@ -550,7 +550,7 @@ mod tests {
 
     /// API sources a witness may live in: the serve crate's unit tests plus
     /// its end-to-end suites.
-    fn api_sources() -> [&'static str; 11] {
+    fn api_sources() -> [&'static str; 13] {
         [
             include_str!("../../stella-serve/src/server.rs"),
             include_str!("../../stella-serve/tests/calibration.rs"),
