+++ stella-serve/src/goal.rs
@@ -34,6 +34,7 @@
 //! turn outcome carrying its real cost — the same contract a cancelled
 //! single-turn run has, which is what lets a host handle one path.
 
+use stella_core::bus::HookBus;
 use stella_core::{BudgetGuard, Engine, EventSender, GoalConfig, TurnOutcome};
 use stella_engine::CancelToken;
 use stella_protocol::{AgentEvent, CompletionMessage, Provider, StageKind};
@@ -113,6 +114,11 @@ pub(crate) async fn drive_goal(
     mut budget: BudgetGuard,
     events: &mpsc::UnboundedSender<AgentEvent>,
     cancel: CancelToken,
+    // This turn's operator hook plane (#1298), forwarded into every round.
+    // A goal run is many turns wearing one turn's id, and an observer that
+    // saw only the first round would report a truncated run — so the bus
+    // reaches each round exactly as it reaches a single served turn.
+    bus: Option<&HookBus>,
 ) -> DrivenTurn {
     let sender = EventSender::new(events.clone());
     messages.push(seed_message(&run.goal));
@@ -121,9 +127,15 @@ pub(crate) async fn drive_goal(
         budget.begin_turn();
         let offset = u32::try_from(2 * round.saturating_sub(1)).unwrap_or(u32::MAX);
         let round_engine = engine.with_turn_instance(base_turn.saturating_add(offset));
-        let driven =
-            crate::session::drive_turn(&round_engine, messages, budget, events, cancel.clone())
-                .await;
+        let driven = crate::session::drive_turn(
+            &round_engine,
+            messages,
+            budget,
+            events,
+            cancel.clone(),
+            bus,
+        )
+        .await;
         messages = driven.messages;
         budget = driven.budget;
         if let TurnOutcome::Aborted { reason, cost_usd } = driven.outcome {
